// packages/element-plus/src/components/crud/src/interface.ts
export interface Scope {
  /** 当前行索引 */
  $index: Index
  /** 当前行数据 */
  row: AnyObject
}
// ...
export type TableOption = {
  /** 表格标题 */
  title?: string | VNode
  /** 表格高度 */
  height?: string
  /** table的最大高度。 合法的值为数字或者单位为 px 的高度。 */
  maxHeight?: string
  /** 是否有序号 默认为`false` */
  index?: boolean
  /** 序号栏宽度 默认为`50` */
  indexWidth?: string | number
  /** 序号栏 文本内容 */
  indexText?: string
  /** 是否懒加载子节点数据(树形表格) */
  lazy?: boolean
  /** 当内容过长被隐藏时显示 tooltip */
  showOverflowTooltip?: boolean
  /** 渲染嵌套数据的配置选项(树形表格) */
  treeProps?: {
    hasChildren?: string
    children?: string
    /** 父节点和子节点的选择状态是否关联 */
    checkStrictly?: boolean
  }
  /** 是否带有纵向边框 默认为`false` */
  border?: boolean
  /**  */
  span?: number
  /** 表单项的文字宽度 默认为`90` */
  labelWidth?: number
  /** 查看弹窗的labelWidth 如果没有取labelWidth 默认为`90` */
  checkLabelWidth?: number
  /** 是否有选择框 默认为`false` */
  selection?: boolean
  /** 多选切换分页是否会情况已选择的数据  请注意， 需指定 rowKey 来让这个功能生效。 默认为`true` */
  reserveSelection?: boolean
  /** 是否有单选列 默认为`false` */
  radio?: boolean
  /** 复选框栏宽度 默认为`50` */
  selectionWidth?: string
  /** 行数据的`Key`，用来优化 `Table`的渲染；在使用 `reserve-selection`功能与显示树形数据时，该属性是必填的。 */
  rowKey?: string
  /** 是否显示操作栏 默认为`false` */
  menu?: boolean
  /** 操作栏标题 */
  menuTitle?: string
  /** 操作栏宽度 默认为`220` */
  menuWidth?: string | number
  /** 操作栏最小宽度 */
  menuMinWidth?: string | number
  /** 操作栏列是否固定在左侧或者右侧。 `true`表示固定在左侧  默认`right`*/
  menuFixed?: boolean | 'left' | 'right'
  /**操作栏表头的对齐方式 */
  menuHeaderAlign?: 'left' | 'center' | 'right'
  /** 操作栏按钮的对齐方式 默认为`center` */
  menuAlign?: 'left' | 'center' | 'right'
  /** 是否开启拖拽排序 默认为`false` */
  sortable?: boolean
  /** 表头配置 */
  column: ColumnProps[]
  /**
   * 表单配置 JSON Schema 格式
   */
  formSchemaField?: SchemaItemArray
  /* column 的 key， 如果需要使用 filter-change 事件，则需要此属性标识是哪个 column 的筛选条件 */
  // filterChange: functionType<(filters: AnyObject) => void>()
  /** 表格等待框的控制 */
  tableLoading?: boolean
  /** 是否要高亮当前行 */
  highlightCurrentRow?: boolean
  /** 搜索项标题宽度 */
  searchLabelWidth?: number
  /** 搜索项框栅列 */
  searchSpan?: number
  /** 是否默认展开所有行，当前表格包含展开行存在或者为树形表格时有效 */
  defaultExpandAll?: boolean
  /** 弹窗是否为抽屉 */
  isDrawer?: boolean
  /** 表格新增弹窗标题 */
  addTitle?: string
  /** 表格修改弹窗标题 */
  editTitle?: string
  /** 表格查看弹窗标题 */
  viewTitle?: string
  /** 表格弹窗宽度 */
  dialogWidth?: string
  /** 表格新增按钮 */
  addBtn?: boolean
  /** 表格删除按钮 */
  delBtn?: boolean
  /** 表格查看按钮 */
  viewBtn?: boolean
  /** 表格修改按钮 */
  updateBtn?: boolean
  /** 表格导入按钮 */
  importBtn?: boolean
  /** 表格导出按钮 */
  excelBtn?: boolean
  /** 表格头部菜单栏右边是否显示 */
  menuHeaderRight?: boolean
  /** 是否卡片模式 */
  isCard?: boolean
  /** 表单操作栏是否显示 */
  menuBtn?: boolean
  /** 表格弹窗取消按钮 */
  cancelBtn?: boolean
  /** 表格弹窗取消按钮文本 */
  cancelBtnText?: string
  /** 表格弹窗提交按钮 */
  submitBtn?: boolean
  /** 表格弹窗提交按钮文本 */
  submitBtnText?: string
  /** 表单查看模式的一行 Descriptions-Item的数量 默认为`2` */
  checkColumnSpan?: number
}

export type AlignType = 'center' | 'left' | 'right'
// export type TableOption = Partial<ExtractPropTypes<ReturnType<typeof tableOption>>>;

export interface CrudPageProps {
  /** 当前页数 */
  currentPage?: number
  /** 每页显示条目个数 */
  pageSize?: number
  /** 总条目数 */
  total?: number
  /** 页码按钮的数量，当总页数超过该值时会折叠 */
  pagerCount?: number
  /** 是否为分页按钮添加背景色 */
  background?: boolean
  /** 组件布局，子组件名用逗号分隔 */
  layout?: Array<'prev' | 'pager' | 'next' | 'jumper' | '->' | 'sizes' | 'total' | 'slot'>
  /** 每页显示个数选择器的选项设置 */
  pageSizes?: number[]
}
// ...
export type PermissionMenBtnType = boolean | ComputedRef<boolean> | ((porps: Scope) => boolean)

/* 权限配置 */
export interface CrudPermission {
  /** 新增按钮权限 */
  addBtn?: boolean | ComputedRef<boolean>
  /** 编辑按钮权限 */
  editBtn?: PermissionMenBtnType
  /** 删除按钮权限 */
  delBtn?: PermissionMenBtnType
  /** 查看按钮权限 */
  viewBtn?: PermissionMenBtnType
  /** 导出按钮权限 */
  exportBtn?: ComputedRef<boolean>
  /** 导入按钮权限 */
  importBtn?: ComputedRef<boolean>
  /** 查询按钮权限 */
  queryBtn?: ComputedRef<boolean>
}